third_party/blink/web_tests/wpt_internal/overscroll-gestures/parsing/overscroll-area-valid.html

<!DOCTYPE html>
<meta charset="utf-8">

<title>Overscroll Area Parsing Test - valid</title>
<meta name="assert" content="overscroll-area supports the full grammar 'none | <dashed-ident>#'.">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<script>
test_valid_value("overscroll-area", "none", "none");
test_valid_value("overscroll-area", "nOnE", "none");
test_valid_value("overscroll-area", "--foo", "--foo");
test_valid_value("overscroll-area", "--foo, --bar", "--foo, --bar");
test_valid_value("overscroll-area",
                 "--foo, --bar, --baz, --bar, --baq",
                 "--foo, --bar, --baz, --baq");
</script>
